Move the renderPoint to PointController
