+++ src/controllers/trip.js
@@ -2,49 +2,19 @@ import SortComponent, { SortType } from '../components/sort';
 import DayListComponent from '../components/day-list';
 import DayComponent from '../components/day';
 import SortEventsComponent from '../components/sort-events';
+import PointController from './point.js';
 
 import NoEventsMessageComponent from '../components/no-events-message';
 
 import { getDate, getUnique } from '../utils/common';
 
-import EventComponent from '../components/event';
-import EventEditComponent from '../components/event-edit';
-
-import { renderComponent, replace } from '../utils/render';
+import { renderComponent } from '../utils/render';
 
 const sortByDurationInDescendingOrder = (a, b) => (b.endTime - b.startTime) - (a.endTime - a.startTime);
 
 const sortByPriceInDescendingOrder = (a, b) => b.price - a.price;
 
 const sortPurely = (collection, iterate) => collection.slice().sort(iterate);
-
-const renderEvent = (eventListElement, event) => {
-  const onEscKeyDown = (evt) => {
-    const isEscKey = evt.key === `Escape` || evt.key === `Esc`;
-
-    if (isEscKey) {
-      startEventEditing();
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    }
-  };
-
-  const startEventEditing = () => replace(eventComponent, eventEditComponent);
-
-  const stopEventEditing = () => replace(eventEditComponent, eventComponent);
-
-  const eventComponent = new EventComponent(event);
-
-  eventComponent.setEditButtonClickHandler(() => {
-    stopEventEditing();
-    document.addEventListener(`keydown`, onEscKeyDown);
-  });
-
-  const eventEditComponent = new EventEditComponent(event);
-
-  eventEditComponent.setSubmitHandler(startEventEditing);
-
-  renderComponent(eventListElement, eventComponent);
-};
 
 const renderDays = (dayListElement, events) => {
   const days = getUnique(events.map((event) => getDate(event.startTime)));
@@ -55,7 +25,10 @@ const renderDays = (dayListElement, events) => {
     const eventList = dayComponent.getElement().querySelector(`.js-trip-events__list`);
     events
       .filter((event) => getDate(event.startTime) === day)
-      .forEach((event) => renderEvent(eventList, event));
+      .forEach((event) => {
+        const pointController = new PointController(eventList);
+        pointController.render(event);
+      });
   });
 };
 
@@ -63,8 +36,10 @@ const renderSortEvents = (dayListElement, events) => {
   const sortEventsComponent = new SortEventsComponent();
   renderComponent(dayListElement, sortEventsComponent);
   const eventList = sortEventsComponent.getElement().querySelector(`.js-trip-events__list`);
-  events
-    .forEach((event) => renderEvent(eventList, event));
+  events.forEach((event) => {
+    const pointController = new PointController(eventList);
+    pointController.render(event);
+  });
 };
 
 export default class TripController {
